requirements.txt
aiosqlite==0.17.0
anyio==3.4.0
asgiref==3.4.1
autopep8==1.6.0
click==8.0.3
colorama==0.4.4
fastapi==0.70.0
h11==0.12.0
idna==3.3
iso8601==0.1.16
pycodestyle==2.8.0
pydantic==1.8.2
pypika-tortoise==0.1.1
pytz==2021.3
sniffio==1.2.0
starlette==0.16.0
toml==0.10.2
tortoise-orm==0.17.8
typing_extensions==4.0.1
uvicorn==0.15.0
